% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annex.R, R/stats.R
\name{annex}
\alias{annex}
\alias{head.annex}
\alias{tail.annex}
\alias{subset.annex}
\alias{head.annex_stats}
\alias{tail.annex_stats}
\alias{subset.annex_stats}
\title{Annex Creator Function}
\usage{
annex(formula, data, tz, duplicate.action = NULL, meta = NULL, verbose = FALSE)

\method{head}{annex}(x, ...)

\method{tail}{annex}(x, ...)

\method{subset}{annex}(x, ...)

\method{head}{annex_stats}(x, ...)

\method{tail}{annex_stats}(x, ...)

\method{subset}{annex_stats}(x, ...)
}
\arguments{
\item{formula}{the formula to specify how the data set is set up.
See 'Details' for more information.}

\item{data}{\code{data.frame} containing the obervations/data.}

\item{tz}{character, time zone definition (e.g., \code{"Europe/Berlin"} or
\code{"UTC"}); required.  \code{OlsonNames()} returns a list of possible
time zones. The correct time zone is important to properly calculate month and time of day.}

\item{duplicate.action}{\code{NULL} or a function which returns
a single numeric value. Used to handle possible duplicates, see
'Details'.}

\item{meta}{\code{NULL} (default) or a \code{list} with information
about study, home, and room (see section 'Duplicates').}

\item{verbose}{logical, defaults to \code{FALSE}. Can be set
to \code{TRUE} to increase verbosity.}

\item{x}{object of class \code{annex}.}

\item{\dots}{arguments to be passed to or from other methods.}
}
\description{
Creates an object of class \code{c("annex", "data.frame")} required to calculate
the statistics.
}
\details{
In case the data set provided on \code{data} does only contain data
of one study, home, and room, the fomula has two parts, looking
e.g., as follows:

\itemize{
\item \code{T + RH ~ datetime}
}

The left hand side of the formula (left of \code{~}) specifies
the names of the variables of the observations to be processed,
the right hand side is the name of the variable containing the
time information (must be of class \code{POSIXt}). In this case,
the \code{meta} argument is required to provide information about
the study, home, and room.

If the grouping information is already in the data set,
the analysis can be performed depending on the group information,
typically:

\itemize{
\item \code{T + RH ~ datetime | study + home + room}
}

The latter allows to process observations from different studies, homes, and/or rooms
all in one go.
}
\section{Duplicates}{


Duplicated records can distort the statistics and should be handled properly.
For each unique \code{study, home, room} only one observation (row) for a specific
date and time should exist.

As there is no general way to deal with such duplicates, the function \code{annex}
(as well as \code{annex_prepare}) by default throws a warning for the user if such
duplicates exist (\code{duplicate.action = NULL}; default argument).

However, the package allows the user to provide a custom \code{duplicate.action}
function, e.g., \code{mean}, \code{min}, \code{max}, ... This function must return
one single numeric value (or an NA) when applied to a vector. If a function is provided,
the \code{annex} function does the following:
\itemize{
\item Checks if there are any duplicates. If not, no changes are made. Else ...
\item Checking if the function is valid (returns single numeric or NA). If not,
an error will be thrown.
\item Takes the measurements of each duplicate; if all values are missing,
an \code{NA} will be returned. Else the users \code{duplicate.action} is applied
to all remaining non-missing values. I.e., if \code{duplicate.action = mean}
the average of all non-missing values will be used.
}
}

\examples{
# Create artificial data set for testing; typically this information is read
# from a file or any other data connection.
data <- data.frame(datetime = as.POSIXct("2022-01-01 00:00", tz = "Europe/Berlin") + -10:10 * 3600,
                   T  = round(rnorm(21, mean = 20, sd = 2), 2),
                   RH = round(runif(21, 40, 100), 2))

res1 <- annex(T + RH ~ datetime, data = data,
              meta = list(study = "example", home = "ex", room = "BED"),
              tz = "Europe/Berlin")
head(res1, n = 3)

# The meta information can also be added to `data` removing the need
# to specify the `meta` argument and allows to mix data from different
# studies and rooms. Appending study, room, and home to `data`:
data <- transform(data,
                  study = "example",
                  home  = "ex",
                  room  = "BED")
head(data)
res2 <- annex(T + RH ~ datetime | study + home + room,
              data = data, tz = "Europe/Berlin")
head(res2, n = 3)

}
\seealso{
annex_prepare, annex_stats
}
\author{
Reto Stauffer
}
