% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annex.R
\name{annex_add_month_and_tod}
\alias{annex_add_month_and_tod}
\title{Calculate month and time of day}
\usage{
annex_add_month_and_tod(x, tz)
}
\arguments{
\item{x}{object of class \code{POSIXt}.}

\item{tz}{time zone (character). Important to properly calculate
month and time of day.}
}
\value{
List with two elements \code{month} (factor) and \code{tod} (factor).
}
\description{
Calculates month and the time of day categories based
on input argument \code{x} with respect to the time zone
specified by the user.
}
\author{
Reto Stauffer
}
