% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annex.R
\name{annex_add_season_and_tod}
\alias{annex_add_season_and_tod}
\title{Calculate season and time of day}
\usage{
annex_add_season_and_tod(x)
}
\arguments{
\item{x}{object of class \code{POSIXt}.}
}
\value{
List with two elements \code{season} (factor) and \code{tod} (factor).
}
\description{
Calculates season and the time of day categories based
on input argument \code{x}.
}
\author{
Reto Stauffer
TODO(R): TIME ZONE HANDLING
}
