% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{annex_check_config}
\alias{annex_check_config}
\title{Checking Annex Config}
\usage{
annex_check_config(x)
}
\arguments{
\item{x}{object of class \code{data.frame} to be checked.}
}
\value{
Invisibly returns a (possibly modified) version of \code{x}
containing only the required columns in a specific order.
}
\description{
An 'annex config object' is a simple data.frame which
can be created using Rs standard features. The function
is checking if the content of the object matches the
requirements of being an annex config object.
This check will be performed in \code{link{annex()}} automatically,
but can also be done by the user manually.
}
\details{
The function checks if the config object is set up properly
and contains the required information for preparing the annex data.

Fails if:
\itemize{
\item TODO(R): checks changed
\item the input is not a \code{data.frame}
\item variables are missing ('column', 'variable', 'study', 'home', 'room')
\item configuration for \code{variable = "datetime"} is missing
\item there is no definition for variables (datetime only)
\item the config contains missing values in the required variables
\item column name for \code{variable = "datetime"} is missing
\item the variable 'column' is not unique
\item there are duplicated entries for the combination of
variable/study/home/room (must be unique)
\item variables study, home or room contain contain non-allowed characters.
Must only contain letters (lowercase or uppercase), numbers,
and underscores. Must start with a letter.
}
}
\author{
Reto Stauffer
}
