% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stats.R
\name{annex_check_stats_object}
\alias{annex_check_stats_object}
\title{Checking Annex Stats to XML}
\usage{
annex_check_stats_object(x)
}
\arguments{
\item{x}{object of class \code{annex_stat} as returned by
\code{\link[=annex_stats]{annex_stats()}}.}
}
\value{
No return, will throw an error if something does
not match the expected file format.
}
\description{
Used for checking/validating \code{annex_stat} objects;
used internally by \code{\link[=annex_write_stats]{annex_write_stats()}} to ensure that
what the users (try to) store to the final XLSX files
is what is expected for the final output file.
}
\details{
The following will be checked:

\itemize{
\item Input is of correct type and has at least one observation.
\item
\item
\item
\item
}
}
\author{
Reto Stauffer
}
