% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISO3166.R
\name{annex_countries}
\alias{annex_countries}
\title{Country codes (ISO 3166) alpha-2 and alpha-3}
\usage{
annex_countries()
}
\value{
Returns the data set \code{ISO3166} shipped
with the package (see \code{?ISO3166} for details).
}
\description{
Country codes (ISO 3166) alpha-2 and alpha-3
}
\author{
Reto Stauffer
}
