% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{annex_dist}
\alias{annex_dist}
\title{Get (Mixed) Empirical Distribution}
\usage{
annex_dist(x, ..., verbose = TRUE)
}
\arguments{
\item{x}{a named vector or \code{data.frame} with empirical quantiles and
sample size (optional). See Section 'Details' for more information.}

\item{verbose}{Logical, if \code{TRUE} some additional output is shown.}

\item{`...`}{see Section 'Details'.}
}
\description{
Get (Mixed) Empirical Distribution
}
\details{
This function returns an \code{annex_edist} (annex empirical distribution) object
which can be based on a single empirical distribution, or a (weighted) mixed
distribution. \code{annex_stats} stores the empirical distribution of our measurements
given a series of quantiles as well as the sample size the quantiles are based on.
This function allows a series of different inputs.

\strong{Named numeric vector:} \code{x} can be a named numeric vector. In this case,
all the names of the vector must be unique and follow
\verb{^p[0-9]\{1,3\}(\\.[0-9]\{,2\})?$}. Some examples: \code{p00} (minimum), \code{p2} (2_th_
percentile or 0.02 quantile), \code{p02.5} (2.5_th_ percentile or 0.025
quantile), \code{p05} (5_th_ percentile or quantile 0.05) etc. up to \code{p100}
(maximum). Minimum (\code{p00}) and maximum (\code{p100}) \emph{must} be given, the
quantiles in between are handled flexible.
}
\author{
Reto Stauffer
}
