% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{annex_dist_process_input}
\alias{annex_dist_process_input}
\title{Helper Function to Prepare User Input on \code{annex_dist}}
\usage{
annex_dist_process_input(x, ..., verbose = TRUE)
}
\value{
Returns prepared (standardized) list for further processing.
}
\description{
Helper function handling the different input types allowed when calling
\code{annex_process_input}. Check man page for \code{annex_process_input} for details.
}
\author{
Reto Stauffer
}
