% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annex.R
\name{annex_handle_duplicates}
\alias{annex_handle_duplicates}
\title{Internal Function for Handling Possible Duplicates}
\usage{
annex_handle_duplicates(x, formula, duplicate.action, verbose = FALSE)
}
\arguments{
\item{x}{data.frame to be processed.}

\item{formula}{object of class \code{Formula}, the formula provided to \code{annex()}.}

\item{duplicate.action}{can be \code{NULL} (no dedicated handling for duplicates)
of a function. If function it must return a single numeric value,
will be tested to be able to provide a useful error for the user if needed.}

\item{verbose}{logical, verbosity (defaults to \code{FALSE}).}
}
\value{
Returns a \code{data.frame} similar to argument \code{x} with possibly
modified content (depending on how to deal with duplicates if any).
}
\description{
This function is called inside \code{annex()}.
}
\author{
Reto
}
