% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stats.R
\name{annex_read_stats}
\alias{annex_read_stats}
\title{Reading Annex Statistics from XLSX}
\usage{
annex_read_stats(file, raw = FALSE, validate = TRUE)
}
\arguments{
\item{file}{character, name of the file(s) to be imported.
Must end on XLSX (not case sensitive).}

\item{raw}{logical. If \code{FALSE} a single \code{data.frame}
will be returned, containing the statistics from
all file(s). If set \code{TRUE} the raw information
of the XLSX file(s) is returned (see Details).}

\item{validate}{logical, if \code{TRUE} (default) the function
first validates if the XLSX file is valid given the current
version of annex.}
}
\value{
An object of class \code{c("annex_xlsx_stats", "data.frame")}
or a named list of \code{data.frame}s. Depends argument \code{raw},
see section 'Details'.
}
\description{
The final data set (annex statistics) are written into an XLSX
file. This function allows to read one or multiple of these files
into R.
}
\details{
The XLSX files (if valid) contain a series of sheets,
namely 'STATS', 'META-Study', 'META-Home', 'META-Room', and
'META-Variable' containing data information. This function
has two different returns (depending on \code{combine}).

If \code{raw = FALSE} (default) only the 'STAT' sheet is
imported. If there is more than one \code{file} the information
from all files will be combined in one \code{data.frame} which
is returned.

If \code{raw = TRUE} all the information from all sheets is
read and stored in a named list. If there are multiple files,
the content of the different sheets are combined. The return
is a named list of length 5 containing a \code{data.frame} each.
For usability purposes the sheet names will be modified,
replacing "-" (XLSX) with "_" (names of list).
}
\author{
Reto Stauffer
}
