% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{annex_room_definition}
\alias{annex_room_definition}
\title{Room definition information}
\usage{
annex_room_definition()
}
\value{
\code{data.frame} with base room abbreviation, long description,
plus a series of examples of valid room labels.
}
\description{
The template contains a series of base abbreviations allowed to define
a room alongside the 'long name'. This function returns the definition
as a \code{data.frame}.
}
\seealso{
annex_variable_definition, annex_room_definition, annex_country_definition
}
\author{
Reto Stauffer
}
