% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{annex_stats}
\alias{annex_stats}
\title{Calculate Statistics on Annex object}
\usage{
annex_stats(object, format = "wide", ..., probs = NULL)
}
\arguments{
\item{object}{an object of class \code{annex}.}

\item{format}{character, either \code{"wide"} (default) or \code{"long"}.}

\item{\dots}{currently unused.}

\item{probs}{\code{NULL} (default; see Details) or a numeric vector of probabilities
with values in \code{[0,1]} (Values will be rounded to closest 3 digits).}
}
\value{
Returns an object of class \code{c("annex_stats", "data_frame")}.
}
\description{
Calculate Statistics on Annex object
}
\details{
The function allows to return the statistics in a wide format or long format.
Both can be used when calling \code{\link[=annex_write_stats]{annex_write_stats()}}, but he long/wide
format can be handy fur custom applications (e.g., plotting, ...).

Argument \code{probs} will be forwarded to the \code{\link[stats:quantile]{stats::quantile()}} function.
If \code{probs = NULL} (default) the empirical quantiles will be calculated
from \code{0} (the minimum) up to \code{1} (the maximum) in an interval of
\code{0.01} (one percent steps), including quantiles \code{0.005},
\code{0.025}, \code{0.975} and \code{0.995}. Can be specified differently
by the user if needed, however, this no longer yields the standard statistics
and the validation will report a problem.
}
\section{Statistics}{


\strong{Grouping:} Statistics are calculated on different subsets (or groups),
typically \code{study}, \code{home}, \code{room}, \code{year}, \code{month},
\code{tod} (time of day).  However, this set can vary depending on the users
function call to \code{annex} (see argument \code{formula}).

\code{annex_stats} calculates a series of data/quality flags as well as statistical
measures.

\strong{Quality:} \code{quality_lower} and \code{quality_upper} contain the fraction of
observations (in percent) falling below the lower and upper defined threshold
(see \code{annex_variable_definition}).
\code{quality_start} and \code{quality_end} contain the day (date only)
where the first non-missing observation was given for the current group; used to
estimate \code{Nestim} (see below).

\strong{Interval:} Time increments of all non-missing observations are calculated in seconds.
The \code{interval_} columns show the five digit summary plus the arithmetic mean of these
intervals. \code{interval_Median} is used to calculate estimate \code{Nestim} (see below).

\strong{Nestim:} Number of estimated observations (see section below)
\strong{N:} Number of non-missing observations
\strong{NAs:} Number of missing observations (\code{NA} in the data set)
\strong{Mean:} \deqn{\bar{x} = \frac{1}{N} \sum_{i = 1}^N x_i}{x_bar = mean(x)} (arithmetic mean)
\strong{Sd:} \deqn{\text{sd}(x) = \sqrt{\frac{1}{N - 1} \sum_{i = 1}^N \big( (x_i - \bar{x})^2\big)}}{sd(x)}
\strong{p:} Probabilites for different quantiles. \code{p00} represents the overall minimum,
\code{p50} the median, \code{p100} the overall maximum of all non-missing values. Uses
the empirical quantile function with \code{type = 7} (default; see \code{quantile}).
}

\section{Estimated number of observations}{


TODO(R): Explain
}

\seealso{
annex_stats_reshape annes_write_stats
}
\author{
Reto Stauffer
}
