% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{annex_stats_reshape}
\alias{annex_stats_reshape}
\title{Reshaping Annex Stats Objects}
\usage{
annex_stats_reshape(x, format = NULL)
}
\arguments{
\item{x}{object of class \code{annex_stats} as returned
by \code{\link[=annex_stats]{annex_stats()}}.}

\item{format}{\code{NULL} by default or one of \code{"long"}
or \code{"wide"} (see Details).}
}
\value{
Returns a reshaped version of the input. If the
Object provided on \code{x} inherits \code{annex_stats_wide} (wide format)
the long format will be returned and vice versa if \code{format = NULL}.
If the format is specified as either \code{"long"} or \code{"wide"}
the long or wide format will be returned (possibly an unmodified version
of the input if the input is already in the desired format).
}
\description{
Reshaping Annex Stats Objects
}
\seealso{
annex_stats
}
\author{
Reto Stauffer
}
