% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{annex_validate}
\alias{annex_validate}
\title{Validate annex Output File}
\usage{
annex_validate(file, user, ...)
}
\arguments{
\item{file}{name of the file to be validated (XLSX file).}

\item{user}{positive integer, the user identifier given
by the project team.}

\item{\dots}{currently unused.}
}
\value{
Some checks will cause an error and stop execution.
Others will cause a message with some information on what
has to be fixed to make the document valid. If the function
does not stop due to an error it will return \code{TRUE}
if the file has been validated as proper, or \code{FALSE}
if issues have been found.
}
\description{
Validate XLSX file created by \code{\link[=annex_write_stats]{annex_write_stats()}}.
Checks if all required sheets/columns are available and that
all user-modified META information has been entered correctly.
}
\author{
Reto Stauffer
}
