% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{check_for_allowed_pollutants}
\alias{check_for_allowed_pollutants}
\title{Checking for Allowed Pollutants}
\usage{
check_for_allowed_pollutants(x)
}
\arguments{
\item{x}{character vector with variable names.}
}
\value{
Character vector (with possibly adjusted) variable
names, or fails.
}
\description{
The XLSX file template.xlsx contains a series of
pre-defined names for the pollutants (sheet 'Definitions').
This function checks if all user defined variable names
are valid. Not case sensitive; will be adjusted if needed.
}
\author{
Reto Stauffer
}
