% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{check_for_allowed_rooms}
\alias{check_for_allowed_rooms}
\title{Checking for Allowed Rooms}
\usage{
check_for_allowed_rooms(x)
}
\arguments{
\item{x}{character vector with room names.}
}
\value{
Character vector (with possibly adjusted) room
names, or fails. Not case sensitive for checking, but
will return everything in uppercase (\code{toupper(x)}).
}
\description{
The XLSX file template.xlsx contains a series of
pre-defined names for the rooms (sheet 'Definitions').
This function checks if all user defined room names
are valid. Not case sensitive; will be adjusted if needed.
}
\seealso{
annex_variable_definition, annex_room_definition, annex_country_definition
}
\author{
Reto Stauffer
}
