% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{check_for_allowed_variables}
\alias{check_for_allowed_variables}
\title{Checking for Allowed Variables}
\usage{
check_for_allowed_variables(x)
}
\arguments{
\item{x}{character vector with variable names.}
}
\value{
Character vector (with possibly adjusted) variable
names, or fails.
}
\description{
The XLSX file template.xlsx contains a series of
pre-defined names for the variables (sheet 'Definitions').
This function checks if all user defined variable names
are valid. Not case sensitive; will be adjusted if needed.
}
\seealso{
annex_variable_definition, annex_room_definition, annex_country_definition
}
\author{
Reto Stauffer
}
