% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_Bedroom}
\alias{demo_Bedroom}
\title{Demo data set Bedroom}
\description{
One of the demo data sets used for testing and for the documentation/manuals.
At tabular text file (CSV alike) containing measurement data.
The data set \code{demo_Bedroom_config} contains the corresponding
configuration used when calling \code{\link[=annex_prepare]{annex_prepare()}}.
}
\author{
Reto Stauffer
}
\keyword{data}
