% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_Bedroom_config}
\alias{demo_Bedroom_config}
\title{Dummy text 4}
\description{
One of the demo data sets used for testing and for the documentation/manuals.
A tabular text file (CSV alike) containing the config for the
\code{demo_Bedroom} data set, which contains the actual measurements.
}
\author{
Reto Stauffer
}
\keyword{data}
