% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_UIBK}
\alias{demo_UIBK}
\title{Demo data set UIBK (XLSX)}
\description{
One of the demo data sets used for testing and for the documentation/manuals.
An XLSX file containing both, the measurements as well as a configuration to
prepare the annex objects.
}
\author{
Reto Stauffer
}
\keyword{data}
