% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Generates evaluation statistic for data frame}
\usage{
evaluate(data_frame, formula, path)
}
\arguments{
\item{data_frame}{a data frame in specific form \code{\link[=formatting]{formatting()}}}

\item{formula}{a formula describing the specific layout of the data frame}

\item{path}{a path with the file name for the generated .csv file}
}
\value{
Returns summary statistics in a convenient form
}
\description{
Creates a statistic containing \code{mean}, \code{sd}, etc. of the
data frame provided
}
\details{
The data frame must be formatted with \code{\link[=formatting]{formatting()}}
beforehand. Depending on the number of levels the size of the resulting
statistic varies.

The data set is successively broken down into user, study, home, room, and
pollutant id and then summarized using statistical variables.
The statistical parameters used are contained in \code{\link[=statistics]{statistics()}}.

The output is saved in .csv format.
}
\seealso{
\code{\link[=formatting]{formatting()}}, \code{\link[base:data.frame]{base::data.frame()}}
}
