% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{formatting}
\alias{formatting}
\title{Formats a data frame in specific format}
\usage{
formatting(
  data_frame,
  formula,
  tz = "UTC",
  format = "\%Y-\%m-\%d \%H:\%M:\%S",
  user = "",
  study = "",
  home = "",
  room = ""
)
}
\arguments{
\item{data_frame}{a data frame}

\item{formula}{a formula describing the specific layout of the data frame}

\item{tz}{a time zone like "UTC", "MET", ...}

\item{format}{the format in which the time stamps are given in the data frame}

\item{user}{abbreviation for the name of the respective scientist}

\item{study}{a natural number (entered as a character)}

\item{home}{a natural number (entered as a character)}

\item{room}{abbreviation for the name of the respective room}
}
\value{
Returns a data frame that has a specific format according to the
formula provided
}
\description{
Generates a new data frame according to the formula provided.
}
\details{
The function adds columns referring to the specific season and
time of day (tod) for each time stamp. Pollutants will be on the left side,
the date terms on the right including season and tod as well as other
mandatory variables (home, room, etc.). If additional variables were given,
these can also be found in the output.

If a mandatory variable is not present in the output, a message is issued.
Furthermore, the evaluate function of ADA cannot be used unless all the required
variables are available. If the formula passed contains more than just the
mandatory variables, the evaluate function of ADA can be used without any
problems.

The output contains at least seven columns in addition to the columns for
the respective pollutants.
}
\seealso{
\code{\link[base:data.frame]{base::data.frame()}}, \code{\link[stats:formula]{stats::formula()}}
}
