% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{get_ID_info}
\alias{get_ID_info}
\title{Get ID information}
\usage{
get_ID_info(x, n = 5L, prefix = " ID")
}
\arguments{
\item{x}{character vector.}

\item{n}{integer, defaults to \code{5}. Number of rows to be explicitly listed (if any).}

\item{prefix}{character or \code{NULL}, default \code{"ID"}.}
}
\value{
Character string with the information where suspicious values have been
found in the XLSX file. If there are no suspicious rows, \code{FALSE} is returned.
}
\description{
Used for warning and error messages. Returns a printable
string with IDs which have been identified.
}
\author{
Reto Stauffer
}
