% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{get_required_columns}
\alias{get_required_columns}
\title{Get required columns for warnings and infos}
\usage{
get_required_columns(sheet)
}
\arguments{
\item{sheet}{character, name of the XLSX sheet.}
}
\value{
Returns \code{NULL} if there is no definition/no required columns,
or a character vector with the exact column name as used in XLSX.
}
\description{
Different sheets contain different columns which require having the user
to specify meta information or other details (no empty cells allowed).
This function returns the names of these columns (as in the XLSX file)
used for validation. If no definition is available, \code{NULL} is returned.
}
\author{
Reto Stauffer
}
