% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{get_row_info}
\alias{get_row_info}
\title{Get row information}
\usage{
get_row_info(x, n = 5L, offset = 1L, prefix = "  Row")
}
\arguments{
\item{x}{integer or logical, see 'Details'.}

\item{n}{integer, defaults to \code{5}. Number of rows to be explicitly listed (if any).}

\item{offset}{integer, number of additional header lines in the XLSX sheet to correct
the row indicator (defaults to \code{1}).}

\item{prefix}{character or \code{NULL}, default \code{"  Row"}.}
}
\value{
Character string with the information where suspicious values have been
found in the XLSX file. If there are no suspicious rows, \code{FALSE} is returned.
}
\description{
Used for warning and error messages. Returns a printable
string with the information where to find these missing, invalid, ... values
in the XLSX file.
}
\details{
If the input is of class integer it is assumed that the values correspond to the
observations in a \code{data.frame}. I.e, \code{1} is the first observation in the
\code{data.frame}.

If the input is a logical vector it is assumed that all entries being \code{TRUE}
are suspicious, and that the order corresponds to the observations in a \code{data.frame}.

The return value will point to these suspicous rows but in the context of an XLSX file
which typically has one additional header line (thus the default \code{offset = 1L}).
}
\author{
Reto Stauffer
}
