% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annex.R
\name{is.regular.annex}
\alias{is.regular.annex}
\title{Check Regularity of an Annex Series}
\usage{
\method{is.regular}{annex}(x, strict = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{annex}.}

\item{strict}{logical, defaults to \code{TRUE}. If \code{FALSE},
regularity (but not strict regularity) will be checked.}
}
\value{
Returns a named logical vector where the name is
a combination of the grouping (study, home, room), the
content the result of checking regularity.
}
\description{
\code{is.regular} is a regular function for checking whether a series
of observations has an underlying regularity or is even
strictly regular. Evaulate for each group of an annex object.
}
\author{
Reto stauffer
}
