% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISO3166.R
\docType{data}
\name{ISO3166}
\alias{ISO3166}
\title{Country codes (ISO 3166) alpha-2 and alpha-3}
\value{
Returns a \code{data.frame} containing
the \code{country} name alongside \code{ISO2}
and \code{ISO3} short name (ISO 3116 alpha-2 and alpha-3
standard).
}
\description{
List of countries with ISO-2 and ISO-3 abbrevations.
Used to validate the final XLSX file.
}
\references{
IBAN (2023). COUNTRY CODES ALPHA-2 & ALPHA-3,
\url{https://www.iban.com/country-codes}, accessed 2023-02-11.
}
\author{
Reto Stauffer
}
\keyword{countries}
\keyword{data}
