% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.annex}
\alias{plot.annex}
\title{Standard plot for annex objects}
\usage{
\method{plot}{annex}(x, bygroup = FALSE, start = NULL, end = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{annex}.}

\item{bygroup}{logical, by default the subplots are
build up on different variables. If \code{TRUE},
all varables from one group will be plotted in one subplot.}

\item{start}{the start time of the period of interest.}

\item{end}{the end time of the period of interest.}
}
\description{
TODO(R)
}
\author{
Reto Stauffer
}
