% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_annex.R
\name{prepare_annex}
\alias{prepare_annex}
\title{Prepare Annex Data}
\usage{
prepare_annex(x, config)
}
\arguments{
\item{x}{data.frame, the data itself.}

\item{config}{data.frame, config information (see
\code{link{check_config()}}).}
}
\value{
Prepared \code{data.frame} for further processing with
the annex package.
}
\description{
TODO(R)
}
\author{
Reto Stauffer
}
