% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{statistics}
\alias{statistics}
\title{Generates compact statistical summary for data frame}
\usage{
statistics(data_frame, season = "ALL", tod = "ALL")
}
\arguments{
\item{data_frame}{a data frame in specific form \code{\link[=formatting]{formatting()}}}

\item{season}{a character referring to the specific season}

\item{tod}{a character referring to the specific time of day}
}
\value{
Return statistical summary of the data set
}
\description{
Creates the statistical summary used by \code{\link[=evaluate]{evaluate()}}
}
\details{
Using \code{\link[=evaluate]{evaluate()}}, the data set is suitably divided and then
combined with this function. Grouping is based on user, study, home, room,
season, and tod.
}
