% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annex.R
\name{summary.annex}
\alias{summary.annex}
\title{Annex summary}
\usage{
\method{summary}{annex}(object, ...)
}
\arguments{
\item{object}{an object of class \code{annex}.}

\item{\dots}{currently unused.}
}
\value{
Returns \code{NULL} (invisible).
}
\description{
Numeric summary of an annex object.
}
\author{
Reto stauffer
}
