% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{variable_basename}
\alias{variable_basename}
\title{Get variable basename}
\usage{
variable_basename(x)
}
\arguments{
\item{x}{Name(s) of the variable (e.g., \code{Other_3}), character.}
}
\value{
Character vector of same length of \code{x} with variable basenames.
}
\description{
It is possible that - for some variables - we do allow to occur more
than once. This is defined in the 'Definitions' table of the XLSX file.
E.g., if we allowe \code{Other} to occur 10 times, that allows to define measurements
for \code{Other}, \code{Other_0}, \code{Other_1}, \code{Other_2}, \dots, \code{Other_99}.
}
\details{
This function strips away the \verb{_[0-9]+$} if needed and returns the basename.
Used internally.
}
\author{
Reto Stauffer
}
