% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annex.R, R/stats.R
\name{annex}
\alias{annex}
\alias{head.annex}
\alias{tail.annex}
\alias{subset.annex}
\alias{head.annex_stats}
\alias{tail.annex_stats}
\alias{subset.annex_stats}
\title{Annex Creator Function}
\usage{
annex(formula, data, meta = NULL, verbose = FALSE)

\method{head}{annex}(x, ...)

\method{tail}{annex}(x, ...)

\method{subset}{annex}(x, ...)

\method{head}{annex_stats}(x, ...)

\method{tail}{annex_stats}(x, ...)

\method{subset}{annex_stats}(x, ...)
}
\arguments{
\item{formula}{the formula to specify how the data set is set up.
see 'Details' for more information.}

\item{data}{\code{data.frame} containing the obervations/data.}

\item{meta}{\code{NULL} (default) or a \code{list} with information
about study, home, and room (see 'Details').}

\item{verbose}{logical, defaults to \code{FALSE}. Can be set
to \code{TRUE} to increase verbosity.}

\item{x}{object of class \code{annex}.}

\item{\dots}{arguments to be passed to or from other methods.}
}
\description{
Creates an object of class \code{c("annex", "data.frame")}.
}
\details{
In case the data set provided on \code{x} does only contain data
of one study, home, and room, the fomula has two parts, looking
e.g., as follows:

\itemize{
\item \code{co2 + voc ~ datetime}
}

The left hand side of the formula (left of \code{~}) specifies
the names of the variables of the observations to be processed,
the right hand side is the name of the variable containing the
time information (must be of class \code{POSIXt}). In this case,
the \code{meta} argument is required to provide information about
the study, home, and room. Will be appended as a new variable
\code{ID}.

If the grouping information is already in the data set,
the analysis can be performed depending on the group information.

\itemize{
\item \code{co2 + voc ~ datetime | ID} ... or even ...
\item \code{co2 + voc ~ datetime | study + home + room}
}
}
\author{
Reto Stauffer
}
