% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annex.R
\name{annex_parse_formula}
\alias{annex_parse_formula}
\title{Parsing Formula}
\usage{
annex_parse_formula(f, verbose = FALSE)
}
\arguments{
\item{f}{object of class \code{Formula}.}

\item{verbose}{logical, defaults to \code{FALSE}. Can be set
to \code{TRUE} to increase verbosity.}
}
\description{
Function used to test and parse a formula used
in different functions in the Annex package.
}
\author{
Reto Stauffer
}
