% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.R
\name{annex_prepare}
\alias{annex_prepare}
\title{Prepare Annex Data}
\usage{
annex_prepare(x, config)
}
\arguments{
\item{x}{data.frame, the data itself.}

\item{config}{data.frame, config information (see
\code{\link[=annex_check_config]{annex_check_config()}}).}
}
\value{
Prepared \code{data.frame} for further processing with
the annex package.
}
\description{
TODO(R)
}
\author{
Reto Stauffer
}
