% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annex.R
\name{annex_stats}
\alias{annex_stats}
\title{Calculate Statistics on Annex object}
\usage{
annex_stats(object, format = "wide", ...)

annex_stats(object, format = "wide", ...)

annex_stats(object, format = "wide", ...)
}
\arguments{
\item{object}{an object of class \code{annex}.}

\item{format}{character, either \code{"wide"} (default) or \code{"long"}.}
}
\value{
Returns an object of class \code{c("annex_stats", "data_frame")}.

Returns an object of class \code{c("annex_stats", "data_frame")}.

Returns an object of class \code{c("annex_stats", "data_frame")}.
}
\description{
Calculate Statistics on Annex object

Calculate Statistics on Annex object

Calculate Statistics on Annex object
}
\author{
Reto Stauffer
}
