% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{annex_variable_definition}
\alias{annex_variable_definition}
\title{Variable definition information}
\usage{
annex_variable_definition(as_list = FALSE)
}
\arguments{
\item{as_list}{logical. If \code{FALSE} (default) a \code{data.frame}
will be returned, if \code{TRUE} a list (see Details).}
}
\value{
Returns either a \code{data.frame} or \code{list} of \code{lists}
which contains the allowed (defined) variables.
}
\description{
The template not only contains the definition of the allowed variables,
it also states whether or not additional information is required (or
optional) and an upper and lower bound to be considered 'valid'.
Used for quality flags.
}
\details{
If \code{as_list = TRUE} a list of lists is returned, else
a \code{data.frame}.

\strong{List:} The name of the list corresponds to the name of the variable,
whereas each entry contains a list with a logical flag if additional
information in the META sheet is \code{required} as well as a numeric
\code{lower} and \code{upper} bound which defines in which range
an observation is considered to be valid. Can be \code{NA} if not
specified (both or one of them).

If \code{as_list = FALSE} (default) the same information is returned
as a \code{data.frame} containing the same information.
}
\seealso{
check_for_allowed_variables
}
\author{
Reto Stauffer
}
