% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stats.R
\name{annex_write_stats}
\alias{annex_write_stats}
\title{Writing Annex Stats to Disc}
\usage{
annex_write_stats(x, file, user, overwrite = FALSE, ..., quiet = FALSE)
}
\arguments{
\item{x}{object of class \code{annex_stat} as returned by
\code{\link[=annex_stats]{annex_stats()}} (wide or long format).}

\item{file}{name (or path) to the XLSX file where to store
the data. Must end with \code{xlsx} (not case sensitive).
See 'Details'.}

\item{user}{positive integer, the user identifier given
by the project team. Will be appended to the data set.}

\item{overwrite}{logical, defaults to \code{FALSE}.
See 'Details'.}

\item{\dots}{not yet used.}

\item{quiet}{logical. If set \code{TRUE} messages will be printed.}
}
\value{
No return, creates a new XLSX file (see argument \code{file})
and stores the data, or updates an existing XLSX file (see
argument \code{overwrite}).
}
\description{
TODO(R)
}
\details{
This function is used to write the annex statistics - the final
output - into an XLSX file. The output is based on a template
file shipped with the package with a predefined format.

If the output \code{file} does not exist, the template will be copied
and modified by (i) saving the data into the "STAT" sheet as well
as pre-filling some additional meta sheets which have to be
manually edited/entered by the user.

By default, \code{overwrite = FALSE}. If the output \code{file} already
exists, the function will be terminated. However, it can be set
to \code{TRUE} to allow \code{\link[=annex_write_stats]{annex_write_stats()}} to manipulate/overwrite
the current data in that XLSX file. It tries to preserve all
custom data (TODO(R): not yet implemented).
}
\author{
Reto Stauffer
}
