% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annex.R
\name{plot.annex}
\alias{plot.annex}
\title{Standard plot for annex objects}
\usage{
\method{plot}{annex}(x, bygroup = FALSE, ...)

\method{plot}{annex}(x, bygroup = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{annex}.}

\item{bygroup}{logical, by default the subplots are
build up on different variables. If \code{TRUE},
all varables from one group will be plotted in one subplot.}
}
\description{
TODO(R)

TODO(R)
}
\author{
Reto Stauffer
}
