% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annex.R
\name{plot.annex_stats}
\alias{plot.annex_stats}
\title{Standard plot for annex_stats objects}
\usage{
\method{plot}{annex_stats}(x, ...)
}
\arguments{
\item{x}{an object of class \code{annex}.}
}
\description{
TODO(R): How can one plot this 5d object?
}
\author{
Reto Stauffer
}
