% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annex.R
\name{summary.annex}
\alias{summary.annex}
\title{Annex Summary}
\usage{
\method{summary}{annex}(object, type = "default", ...)
}
\arguments{
\item{object}{an object of class \code{annex}.}

\item{type}{character, one of \code{"default"} (default) or \code{"statistic"}.
If \code{type = "statistics"} the result of \code{annex_stats(object)} will be printed.}

\item{\dots}{currently unused.}
}
\value{
Returns \code{NULL} (invisible).
}
\description{
Numeric summary of an annex object.
}
\author{
Reto stauffer
}
